pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "nltk>=3.9.1",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "trafilatura>=2.0.0",
    "wordninja>=2.0.0",
    "flask-login>=0.6.3",
    "pyenchant>=3.2.2",
    "enchant>=0.0.1",
]
